+++ lib/app_component.html
@@ -1,38 +1,36 @@
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous"> 
-</head>
 <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-        <a class="navbar-brand" href="#">Energized ID</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarColor02" aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation">
-          <span class="navbar-toggler-icon"></span>
-        </button>
-    
-        <div class="collapse navbar-collapse" id="navbarColor02">
-          <ul class="navbar-nav mr-auto">
-            <li class="nav-item active">
-              <a class="nav-link" href="">Production <span class="sr-only">(current)</span></a>
-            </li>
-            <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                      Database
-                    </a>
-                    <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                      <a class="dropdown-item" href="#">Database Editor</a>
-                      <a class="dropdown-item" href="#">Database Field Connector</a>
-                    </div>
-                  </li>
-            <li class="nav-item">
-              <a class="nav-link" href="#">Credits</a>
-            </li>
-          </ul>
-          <form class="form-inline">
-            <button class="btn btn-outline-light my-2 my-sm-0" type="button">Log Out</button>
-          </form>
+  <a class="navbar-brand" href="#">Energized ID</a>
+  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarColor02"
+    aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation">
+    <span class="navbar-toggler-icon"></span>
+  </button>
+
+  <div class="collapse navbar-collapse" id="navbarColor02">
+    <!--TODO highlight active route link-->
+    <ul class="navbar-nav mr-auto">
+      <li class="nav-item">
+        <a class="nav-link" [routerLink]="Routes.idScreen.toUrl()">Production</a>
+      </li>
+      <li class="nav-item dropdown">
+        <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown"
+          aria-haspopup="true" aria-expanded="false">
+          Database
+        </a>
+        <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+          <a class="dropdown-item" href="#">Database Editor</a>
+          <a class="dropdown-item" href="#">Database Field Connector</a>
         </div>
-      </nav>
+      </li>
+      <li class="nav-item">
+        <a class="nav-link" href="#">Credits</a>
+      </li>
+    </ul>
+    <form class="form-inline">
+      <button class="btn btn-outline-light my-2 my-sm-0" type="button">Log Out</button>
+    </form>
+  </div>
+</nav>
 
-<id-form [idBloc]="idBloc"></id-form>
-
-<id-preview [idBloc]="idBloc"></id-preview>
+<main>
+  <router-outlet [routes]="Routes.all"></router-outlet>
+</main>
