+++ lib/src/screens/id_screen/id_screen.html
@@ -1,0 +1,3 @@
+<id-form [idBloc]="idBloc"></id-form>
+
+<id-preview [idBloc]="idBloc"></id-preview>
