+++ pubspec.yaml
@@ -10,6 +10,7 @@ environment:
 dependencies:
   angular: ^5.2.0
   angular_bloc: ^0.9.0
+  angular_router: ^2.0.0-alpha
   http: ^0.12.0
   image: ^2.1.4
