+++ web/main.dart
@@ -1,10 +1,12 @@
 import 'package:angular/angular.dart';
+import 'package:angular_router/angular_router.dart';
 import 'package:energized_id_web/app_component.template.dart' as ng;
 import 'package:http/http.dart' as http;
 
 import 'main.template.dart' as self;
 
 @GenerateInjector([
+  routerProvidersHash,
   ClassProvider(http.Client),
 ])
 final InjectorFactory injector = self.injector$Injector;
